requirements.txt
streamlit
pandas
numpy
plotly
python-dotenv
requests
